Refactor parser
This commit does not change functionality. Some method
signatures have changed, which might affect derived parsers.
We also fix a few lint warnings outside the parser.

Here are the major refactorings:

1. Remove unused imports and variables;
2. Make variables 'final' wherever possible;
2. Remove unused initializers;
4. Convert optional blocks to mandatory blocks. For example,

  [ e = X() ]

becomes

  ( e = X() | { e = null; } )

and now we can remove e's initializer and make it 'final'.

5. Deprecate the 'startList' method. Previously you would
use 'startList' to initialize the list when adding the
first element:

  List<SqlNode> list;
  ...
  list = startList(e)

Now you initialize the list when it is declared:

  final List<SqlNode> list = new ArrayList<SqlNode>();
  ...
  list.add(e);

This is a bit more work if the list is never used. But it
makes the code to add the first element the same as the code
to add subsequent elements.

6. If parser methods are used to build lists, this change
refactors them to pass the list as an argument. For
example,

  SqlNode OrderItem() : ...

  e = OrderItem() { list.add(e); }
  ( <COMMA> e = OrderItem() { list.add(e); } )*

is now

  void AddOrderItem(List<SqlNode> list) : ...

  AddOrderItem(list) ( <COMMA> AddOrderItem(list) )*

This change adds a 'Add' prefix to existing methods of this
style.

7. Remove an obsolete resource that used to be used by the
parser.


Former-commit-id: 515f3356a6a1ab4bd570c1c20bec9a7e5d4aca5f