Various improvements to OperandTypes
In SqlSingleOperandTypeChecker.java add default method
implementations so that subtypes don't need to.

Add OperandTypes.interval(SqlTypeName) to match an interval
operand.

Add CompositeOperandTypeChecker.withGenerator, so that a
checker can generate signatures given a function name, as
opposed to hard-coding the function name in the signatures
as today.

Add SqlOperandTypeChecker.and,
SqlSingleOperandTypeChecker.or, and similar methods, to
combine checkers using a binary operation. Flatten composite
checkers, so that 'or(or(a, b), c)' becomes 'or(a, b,c)'.


Former-commit-id: 33f4ab40bbee26e06209061c35a422f2f1e05371