[CALCITE-5391] JoinOnUniqueToSemiJoinRule should preserve field names, if possible
Why:
When rewriting the query for a semi join, the
`JoinOnUniqueToSemiJoinRule` pushes a new project via
`RelBuilder`. This project is *almost* a good copy of the original
`Project` -- it is just missing the field names.

How:
This change uses `RelDataType` of the `Project` to get the field names
for the new `Project` pushed on the `Relbuilder` stack. This is
similar to what is done in `SemiJoinRule#perform`

Former-commit-id: 725b71e5af13f2b8725a8da9762c21659c52e507