[CALCITE-5417] Upgrade Proj4J from 1.1.5 to 1.2.2, and restore it as an included dependency
In [CALCITE-5399] we removed Proj4J as a dependency due to
licensing issues with the EPSG data set. In this change we
upgrade Proj4J to version 1.2.2, which no longer includes
the EPSG data set and therefore can be distributed under the
Apache license.

This change mostly consists of a reversal of [CALCITE-5399],
but also upgrades Proj4J.


Former-commit-id: 0a7d84270a0e23c1631a2732d41c1a0dd71527f2