[CALCITE-5522] Babel parser cannot handle some overloads of the DATE_TRUNC function
This is a regression caused by [CALCITE-5447]. Adding support
for `DATE_TRUNC(expr, timeUnitName)` caused the parser to no
longer support `DATE_TRUNC(expr, expr)` such as
`DATE_TRUNC('month', orders.order_date)`. The parser saw an
identifier and wrongly concluded that it must be the name of
a custom time frame.

The solution is for the parser to treat `timeUnitName` as
just an expression (that happens to be a `SqlIdentifier`).
The validator can decide that it is a custom time frame and
convert it to a `SqlIntervalQualifier`.

Add interface SqlOperandHandler to allow SqlBasicFunction
to customize how it handles particular arguments, without
needing a new subclass of SqlFunction. The SqlOperandHandler
interface is similar to other strategies (e.g.
SqlReturnTypeInference, SqlOperandTypeChecker) and we think
that it will be useful in solving other problems in future.

In future work, we should remove from the parser specific
treatment of functions, such as TIMESTAMP_TRUNC, that have
one time frame argument.


Former-commit-id: 39e2b22d298237f1f38051d20108bcea6f6f9dc5